git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@142424 4525493e-7705-40b1-a816-d608a930855b
Former-commit-id: 3566d2b5ac33eb2d822c8834eabace3d5e87d1b8

+++ Online/Online/ROMon/src/Applications/HLTFileEqualizer.h
@@ -57,7 +57,7 @@ class HLTFileEqualizer
     DefHltInfoHandler *m_InfoHandler;
     int m_nnodes;
     int m_nfiles;
-    int m_nfiles2;
+    long m_nfiles2;
     int m_low;
     int m_high;
     DimInfo *m_DefStateInfo;
